feat: swap projects to projects_count in projects model

+++ backend/projects/serializers.py
@@ -12,7 +12,7 @@ class TagSerializer(serializers.ModelSerializer):
 class ProjectSerializer(serializers.ModelSerializer):
   id = serializers.CharField(read_only=True)
   total_votes = serializers.IntegerField(read_only=True)
-  votes_average = serializers.IntegerField(read_only=True)
+  votes_average = serializers.FloatField(read_only=True)
   author = serializers.PrimaryKeyRelatedField(read_only=True)
   tags = serializers.StringRelatedField(many=True, read_only=True)
   reviews = ReviewSerializer(many=True, read_only=True)
